search/requirements.txt
PyPDF2
mutagen
python-docx